+++ lightninglabs/index.js
@@ -4,6 +4,7 @@ const decodeSwapRecovery = require('./decode_swap_recovery');
 const encodeSwapRecovery = require('./encode_swap_recovery');
 const getSwapInQuote = require('./get_swap_in_quote');
 const getSwapInTerms = require('./get_swap_in_terms');
+const getSwapMacaroon = require('./get_swap_macaroon');
 const getSwapOutQuote = require('./get_swap_out_quote');
 const getSwapOutTerms = require('./get_swap_out_terms');
 const lightningLabsSwapService = require('./lightning_labs_swap_service');
@@ -18,6 +19,7 @@ module.exports = {
   encodeSwapRecovery,
   getSwapInQuote,
   getSwapInTerms,
+  getSwapMacaroon,
   getSwapOutQuote,
   getSwapOutTerms,
   lightningLabsSwapService,