+++ lightninglabs/service_socket.js
@@ -3,6 +3,7 @@ const {sockets} = require('./conf/swap_service');
 /** Get service socket
 
   {
+    [is_free]: <Use Free Service Endpoint Bool>
     network: <Network Name String>
   }
 
@@ -14,14 +15,23 @@ const {sockets} = require('./conf/swap_service');
     socket: <Service Socket String>
   }
 */
-module.exports = ({network}) => {
-  if (!network) {
+module.exports = args => {
+  if (!args.network) {
     throw new Error('ExpectedNetworkNameForSwapServiceSocket');
   }
 
-  if (!sockets[network]) {
+  if (!!args.is_free && !sockets.free[args.network]) {
     throw new Error('ExpectedKnownNetworkForServiceSocket');
   }
 
-  return {socket: sockets[network]};
+  // Exit early when using the free endpoint
+  if (!!args.is_free) {
+    return {socket: sockets.free[args.network]};
+  }
+
+  if (!sockets.paid[args.network]) {
+    throw new Error('ExpectedKnownNetworkForServiceSocket');
+  }
+
+  return {socket: sockets.paid[args.network]};
 };